% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{seasons}
\alias{seasons}
\alias{year2season}
\alias{date2season}
\title{Seasons convertion functions}
\usage{
year2season(years)

date2season(dates = Sys.Date(), cutoff = "08-01")
}
\arguments{
\item{years}{a vector of integerish numbers, representing the first year of
the season.}

\item{dates}{a vector of dates. Uses today's date by default.}

\item{cutoff}{a single string of the format "%m-%d". Dates before the cutoff
will be identified as the season starting from the previous year.}
}
\value{
Vector of dates of the same length as \code{years}/\code{dates}.
}
\description{
Converts years or dates to season names.
}
\examples{
years <- 2010:2021
year2season(years)

dates <- seq.Date(as.Date("2021-01-01"), as.Date("2022-01-01"), "month")
setNames(date2season(dates), dates)

}
